test(integration): refactor test directory
move integration tests and unit tests under the /tests directory
