+++ jest.integration.config.ts
@@ -4,9 +4,9 @@ const { compilerOptions } = require('./tsconfig.json');
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
-  setupFilesAfterEnv: ['<rootDir>/intTests/setup/setupFile.ts'],
-  roots: ['<rootDir>/intTests'],
-  testPathIgnorePatterns: ['<rootDir>/test'],
+  setupFilesAfterEnv: ['<rootDir>/tests/integration/setup/setupFile.ts'],
+  roots: ['<rootDir>/tests/integration'],
+  testPathIgnorePatterns: ['<rootDir>/tests/unit'],
   transformIgnorePatterns: ['/node_modules/(?!(axios)/).*'],
   transform: {
     '^.+\\.ts?$': 'ts-jest',
