+++ src/config/index.ts
@@ -9,6 +9,7 @@ const {
   BUCKET_NAME,
   CLIENT_EMAIL,
   PRIVATE_KEY,
+  DB_DATABASE,
   DB_MEMORY,
   NODE_ENV,
   PORT,
@@ -34,6 +35,7 @@ export {
   AWS_REGION,
   BUCKET_NAME,
   CLIENT_EMAIL,
+  DB_DATABASE,
   DB_MEMORY,
   LOG_DIR,
   LOG_FORMAT,
