+++ tests/unit/controllers/stripe.controller.test.ts
@@ -1,5 +1,6 @@
-import StripeController from '../../src/controllers/stripe.controller';
-import stripe from '../../src/services/clients/stripe.client';
+import { Request, Response } from 'express';
+import StripeController from '../../../src/controllers/stripe.controller';
+import stripe from '../../../src/services/clients/stripe.client';
 
 describe('Stripe Controller Unit Tests', function () {
   let createMock;
